Update  - Aggiunto l'approccio basato su compressione delle immagini - Aggiunto l'approccio basato sulle wavelet - Pulizia del main e creazione della librerie HoloUtils.py
